Inline main rollup config
